requirements.txt
tkinter
Pillow>=9.0.0
screeninfo>=0.8.1
